backend/requirements.txt
# Core Backend Framework
flask==3.0.0
flask-cors==4.0.0

# AI/ML Libraries
transformers==4.37.0
torch==2.2.0
diffusers==0.25.0
accelerate==0.26.0

# Hugging Face Integration
huggingface-hub==0.20.0
sentencepiece==0.1.99
protobuf==4.25.0

# Image Processing
Pillow==10.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Performance
gunicorn==21.2.0

# Optional: Model Optimization
bitsandbytes==0.42.0
optimum==1.16.0